Organize the package.
Separate out some modules for smaller .py files and clearer dependencies.
